backend/auggietaskmanager/requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
ruff==0.15.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
requests==2.32.5
icalendar==7.0.1
pytz==2025.2
django-stubs==5.2.9
pillow==12.2.0
